requirements.txt
fastapi
uvicorn
pydantic
motor
pytest
pytest-bdd
pytest-asyncio
httpx
python-dotenv
confluent-kafka
easyocr
langchain
pillow
python-multipart==0.0.6
urllib3==1.26.6
langchain-community
openai
numpy 
opencv-python